progress in Meth_emC. Many files are only able to compile
... with #ifdef DEF_ClassJc_Vtbl or if DEF_REFLECTION_FULL is set.
Should be also part of the project, especially the Visual Studio for test
though DEF_REFLECTION_FULL and DEF_ClassJc_Vtbl is not used, and not all feature
of this files are used. mark with conditional compiling.

+++ emC/Ctrl/T1_Ctrl_emC.h
@@ -65,7 +65,7 @@ static inline int16 step_T1i_Ctrl_emC(T1i_Ctrl_emC_s* thiz, int16 x) {
 }
 
 static inline int16 step32_T1i_Ctrl_emC(T1i_Ctrl_emC_s* thiz, int16 x) {
-  thiz->dx.v32 = (int32)(((uint64)(thiz->fTs.v32) * ( (int32)(x<<16) - thiz->q.v32))>>32);
+  thiz->dx.v32 = (int32)(((uint64)(thiz->fTs.v32) * (uint64)( (int32)(x<<16) - thiz->q.v32))>>32);
   thiz->q.v32 += thiz->dx.v32;
   return thiz->q.v16.hi; //hi part 16 bit
 }
